+++ buckaroo/analysis.py
@@ -2,7 +2,8 @@ from functools import reduce
 import pandas as pd
 from pandas.io.json import dumps as pdumps
 import numpy as np
-
+from buckaroo.pluggable_analysis_framework import ColAnalysis
+'''
 class DfStats(object):
     def __init__(self,
             df,
@@ -29,6 +30,7 @@ class DfStats(object):
         except Exception as e:
             print(e)
             self.col_order = self.df.columns
+'''
 
 def probable_datetime(ser):
     s_ser = ser.sample(np.min([len(ser), 500]))
@@ -52,21 +54,21 @@ def get_mode(ser):
 
 
 class TypingStats(ColAnalysis):
-    provides_summary = [
+    provided_summary = [
         'dtype', 'is_numeric', 'is_integer', 'is_datetime',]
 
     @staticmethod
     def summary(sampled_ser, summary_ser, ser):
-        return {
+        return dict(
             dtype=ser.dtype,
             is_numeric=pd.api.types.is_numeric_dtype(ser),
             is_integer=pd.api.types.is_integer_dtype(ser),
             is_datetime=probable_datetime(ser),
             memory_usage=ser.memory_usage()
-            }
+            )
 
 class DefaultSummaryStats(ColAnalysis):
-    provides_summary = [
+    provided_summary = [
         'length', 'min', 'max', 'mean', 'nan_count', 'distinct_count',
         'distinct_per', 'empty_count', 'empty_per', 'unique_per', 'nan_per',
         'mode']
@@ -80,16 +82,16 @@ class DefaultSummaryStats(ColAnalysis):
         unique_count = len(val_counts[val_counts==1])
         empty_count = val_counts.get('', 0)
 
-    return dict(
-        length=l,
-        nan_count = nan_count,
-        distinct_count= distinct_count,
-        distinct_per = distinct_count/l,
-        empty_count = empty_count,
-        empty_per = empty_count/l,
-        unique_per = unique_count/l,
-        nan_per = nan_count/l,
-        mode=get_mode(ser))
+        return dict(
+            length=l,
+            nan_count=nan_count,
+            distinct_count=distinct_count,
+            distinct_per=distinct_count/l,
+            empty_count=empty_count,
+            empty_per=empty_count/l,
+            unique_per=unique_count/l,
+            nan_per=nan_count/l,
+            mode=get_mode(ser))
 
 def summarize_numeric(ser):
